Move old unit test into a legacy folder
